Force new ice-free regions on the high-res grid
If the ice sheet retreats on the model mesh, make sure that the
area covered by that vertex on the fine grid is also rendered
ice-free. This is done by setting a local mesh dHi to a large
negative number there, then mapping and adding that to the
reference ice thickness on the grid, then setting any resulting
negative values to 0.

Compute the new surface elevation on the fine grid using the new
Hb and Hi.