Mesh regeneration via the grid  instead of the mesh (#4)
* Map the deltas of the mesh back to the reference geometry to create the new mesh

This allows the use of only a single mesh creation routine from the mesh, instead of
having to maintain another one for mesh to mesh generation. 

* Use a coarser grid to map the deltas of the mesh to

The calculation of the mapping operator between the reference
geometry and the fine grid is very expensive, so we trade accuracy
for speed in this case.

* Force new ice-free regions on the high-res grid

If the ice sheet retreats on the model mesh, make sure that the
area covered by that vertex on the fine grid is also rendered
ice-free. This is done by setting a local mesh dHi to a large
negative number there, then mapping and adding that to the
reference ice thickness on the grid, then setting any resulting
negative values to 0.

Compute the new surface elevation on the fine grid using the new
Hb and Hi.

* Add sea level to the computation of grid-based new mesh

Remap the modelled sea level as well as part of the mesh regeneration.
Sea level (+ Hi and Hb) is used to determine the surface elevation,
but there could be a mesh update during a time where sea level has changed.

* introduce the old grid2grid remapping routines to go from the coarser grid to the finer grid

For this mapping we use a simpler 1st order conservative approximation. as this grid is only 
used to generate the new mesh, and specifically _not_ to map the new values. that routine
is still mesh2mesh.